+++ cmd/query/app/gqlgen/gqlgen.yml
@@ -30,7 +30,7 @@ resolver:
 # gqlgen will search for any type names in the schema in these go packages
 # if they match it will use them, otherwise it will generate them.
 autobind:
-  - "github.com/jaegertracing/jaeger/cmd/query/app/graphql/graph/model"
+  - "github.com/jaegertracing/jaeger/cmd/query/app/gqlgen/graph/model"
 
 # This section declares type mapping between the GraphQL and go type systems
 #
