+++ cmd/query/app/gqlgen/graph/schema.graphqls
