+++ cmd/query/app/gqlgen/server.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/99designs/gqlgen/graphql/handler"
 	"github.com/99designs/gqlgen/graphql/playground"
-	"github.com/jaegertracing/jaeger/cmd/query/app/graphql/graph"
-	"github.com/jaegertracing/jaeger/cmd/query/app/graphql/graph/generated"
+	"github.com/jaegertracing/jaeger/cmd/query/app/gqlgen/graph"
+	"github.com/jaegertracing/jaeger/cmd/query/app/gqlgen/graph/generated"
 )
 
 const defaultPort = "8080"
